Standardize the builder pattern
