--- java/semantickernel-connectors-parent/semantickernel-connectors/src/main/java/com/microsoft/semantickernel/connectors/ai/openai/textcompletion/OpenAITextCompletionBuilder.java
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-package com.microsoft.semantickernel.connectors.ai.openai.textcompletion;
-
-import com.microsoft.openai.OpenAIAsyncClient;
-import com.microsoft.semantickernel.textcompletion.TextCompletion;
-
-public class OpenAITextCompletionBuilder implements TextCompletion.Builder {
-    @Override
-    public TextCompletion build(OpenAIAsyncClient client, String modelId) {
-        return new OpenAITextCompletion(client, modelId);
-    }
-}
