+++ java/semantickernel-connectors-parent/semantickernel-connectors/src/main/resources/META-INF/services/com.microsoft.semantickernel.ai.embeddings.EmbeddingGeneration$Builder
@@ -1,0 +1,1 @@
+com.microsoft.semantickernel.connectors.ai.openai.textembeddings.OpenAITextEmbeddingGeneration$Builder
