+++ java/semantickernel-connectors-parent/semantickernel-connectors/src/main/resources/META-INF/services/com.microsoft.semantickernel.textcompletion.TextCompletion$Builder
@@ -1,0 +1,1 @@
+com.microsoft.semantickernel.connectors.ai.openai.textcompletion.OpenAITextCompletion$Builder
