--- java/semantickernel-core/src/main/java/com/microsoft/semantickernel/semanticfunctions/DefaultPromptTemplateBuilder.java
@@ -1,10 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-package com.microsoft.semantickernel.semanticfunctions;
-
-public class DefaultPromptTemplateBuilder implements PromptTemplate.Builder {
-
-    @Override
-    public PromptTemplate build(String promptTemplate, PromptTemplateConfig config) {
-        return new DefaultPromptTemplate(promptTemplate, config);
-    }
-}
