+++ java/semantickernel-core/src/main/resources/META-INF/services/com.microsoft.semantickernel.Kernel.InternalBuilder
@@ -1,1 +1,1 @@
-com.microsoft.semantickernel.DefaultKernelBuilder
+com.microsoft.semantickernel.DefaultKernel$Builder
