+++ java/semantickernel-core/src/main/resources/META-INF/services/com.microsoft.semantickernel.semanticfunctions.PromptTemplate$Builder
@@ -1,0 +1,1 @@
+com.microsoft.semantickernel.semanticfunctions.DefaultPromptTemplate$Builder
