--- java/semantickernel-core/src/main/resources/META-INF/services/com.microsoft.semantickernel.textcompletion.TextCompletion.Builder
@@ -1,1 +1,0 @@
-com.microsoft.semantickernel.connectors.ai.openai.textcompletion.OpenAITextCompletionBuilder
